refactor(agents): split pilot.ts into modular components (Issue #697)
Split the 876-line pilot.ts into smaller, more maintainable modules:

| File | Description | Lines |
|------|-------------|-------|
| pilot/types.ts | Type definitions (PilotCallbacks, PilotConfig, MessageData) | ~70 |
| pilot/message-builder.ts | Message building logic (buildEnhancedContent, buildToolsSection) | ~200 |
| pilot/index.ts | Core Pilot class with orchestration logic | ~600 |

Changes:
- Extract MessageBuilder class for enhanced content building
- Extract type definitions to separate module
- Update all import paths in dependent files
- All 1594 tests pass

Benefits:
- Better separation of concerns
- Easier to test individual components
- Improved code maintainability
- Clearer module boundaries

Fixes #697
