+++ data/rest-files/3c46b2b4-57cf-4470-a112-3e0f7bc163da/download-test.txt
@@ -1,0 +1,1 @@
+Download test content
