+++ data/rest-files/b01b9947-7fcf-4d07-b523-0752cdf2daeb/info-test.txt
@@ -1,0 +1,1 @@
+test content
