+++ src/agents/factory.test.ts
@@ -4,7 +4,7 @@
 
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import { AgentFactory } from './factory.js';
-import { Pilot } from './pilot.js';
+import { Pilot } from './pilot/index.js';
 
 // Mock Config module
 vi.mock('../config/index.js', () => ({
