+++ src/agents/pilot/types.ts
@@ -1,0 +1,84 @@
+/**
+ * Type definitions for Pilot agent.
+ *
+ * Extracted from pilot.ts for better separation of concerns (Issue #697).
+ */
+
+import type { ChannelCapabilities } from '../../channels/types.js';
+import type { FileRef } from '../../file-transfer/types.js';
+import type { BaseAgentConfig } from '../base-agent.js';
+
+/**
+ * Callback functions for platform-specific operations.
+ */
+export interface PilotCallbacks {
+  /**
+   * Send a text message to the user.
+   * @param chatId - Platform-specific chat identifier
+   * @param text - Message content
+   * @param parentMessageId - Optional parent message ID for thread replies
+   */
+  sendMessage: (chatId: string, text: string, parentMessageId?: string) => Promise<void>;
+
+  /**
+   * Send an interactive card to the user.
+   * @param chatId - Platform-specific chat identifier
+   * @param card - Card JSON structure
+   * @param description - Optional description for logging
+   * @param parentMessageId - Optional parent message ID for thread replies
+   */
+  sendCard: (chatId: string, card: Record<string, unknown>, description?: string, parentMessageId?: string) => Promise<void>;
+
+  /**
+   * Send a file to the user.
+   * @param chatId - Platform-specific chat identifier
+   * @param filePath - Local file path to send
+   */
+  sendFile: (chatId: string, filePath: string) => Promise<void>;
+
+  /**
+   * Called when the Agent query completes (result message received).
+   * Used to signal completion to communication layer (e.g., REST sync mode).
+   * @param chatId - Platform-specific chat identifier
+   * @param parentMessageId - Optional parent message ID for thread replies
+   */
+  onDone?: (chatId: string, parentMessageId?: string) => Promise<void>;
+
+  /**
+   * Get the capabilities of the channel for a specific chat.
+   * Used for capability-aware prompt generation (Issue #582).
+   * @param chatId - Platform-specific chat identifier
+   * @returns Channel capabilities or undefined if not available
+   */
+  getCapabilities?: (chatId: string) => ChannelCapabilities | undefined;
+}
+
+/**
+ * Configuration options for Pilot.
+ *
+ * Issue #644: Added chatId binding for session isolation.
+ */
+export interface PilotConfig extends BaseAgentConfig {
+  /**
+   * The chatId this Pilot is bound to.
+   * Each Pilot instance serves exactly one chatId.
+   */
+  chatId: string;
+
+  /**
+   * Callback functions for platform-specific operations.
+   */
+  callbacks: PilotCallbacks;
+}
+
+/**
+ * Message data for processing.
+ */
+export interface MessageData {
+  text: string;
+  messageId?: string;
+  senderOpenId?: string;
+  attachments?: FileRef[];
+  /** Chat history context for passive mode (Issue #517) */
+  chatHistoryContext?: string;
+}
